Rename archival to wagtail_deletion_archival
