+++ .env_SAMPLE
@@ -181,4 +181,4 @@ export HUD_API_ENDPOINT=http://localhost:8000/hud-api-replace/
 ###############################
 # Deletion archival storage
 ###############################
-# export ARCHIVE_PATH=/path/to/write/deleted/content/json/files
+# export WAGTAIL_DELETION_ARCHIVE_PATH=/path/to/write/deleted/content/json/files
